+++ CHANGES.rst
@@ -5,6 +5,9 @@
 Unreleased
 ==========
 
+- Added a new "metric" under /metrics endpoint which exposes NodeInfoMBean
+  attributes (NodeId, NodeName).
+
 2018/03/01 0.1.1
 ================